When we don’t find a matching entity in wikidata for an AACT table row, still add it to the lookup table with an empty qcode.  Then when we go to export studies, only consider those with a qcode.
We’re doing this to try to expedite populating the lookup tables.  So we can launch the process to populate the lookup table and it will know not to check on any rows that already have a row in the lookup table.  We can plan to periodically completely refresh a lookup table, since new entities will be continually added to wikidata, and we want to include those that relate.